invocationMessage: BUG: Separation of stdout / stderr / terminal streams is wrong
The user-supplied COMMANDs need to be run with stdout and stderr swapped, as we need to process stderr, and pass-through stdout (to any external command substitution or redirection). To make (swapped) stderr appear as stdout to the outside, we need another swapping around the command execution + parsing loop.
The final output of command errors needs to separate stderr and terminal outputs; though it doesn't matter for the (existing) tests, in reality this leads to wrong output if stderr is redirected. Introduce INVOCATIONMESSAGE_STDERR_TO_TERM variable to allow tests to control this.