invocationNotification: BUG: Separation of stdout / stderr / terminal streams is wrong
The user-supplied COMMANDs need to be run with stdout and stderr swapped, as we need to process stderr, and pass-through stdout (to any external command substitution or redirection). To make (swapped) stderr appear as stdout to the outside, we need another swapping around the command execution + parsing loop.
Tests: Also cover direct capture of stdout and stderr. The run() function provided by Bats always captures both, but we need to ensure the correct API behavior of the different streams, too.