Make --spinner and --sweep pass stderr output, too
This is now a true differentiation between --spinner-stderr -c "COMMAND | tee /dev/stderr" and --spinner. Output is powering the spinner / sweep, but is not lost.
As we have only a single stream to read from, we somehow need to sort stdout and stderr into separate buckets. A sed stage that is inserted into tee marks stdout lines. sed also ensures that only full individual lines are marked. Inside the processing, we need to extract the stderr output and append that to the existing commandErrorOutput, so that it gets collected and printed at the end.
There seems to be a timing issue with the tee pipeline, which could still produce output even after the assumed final status line is printed. To avoid that, a "sleep 0" seems to do the trick. (Waiting via "wait" did not.)