invocationNotification: Clear the rendering after all lines have been read
Keeping the last rendered state is confusing and was an oversight; invocationMessage also clears that. When no sigil is printed and the message isn't cleared, it needs to be explicitly removed. (But we'd like to avoid superfluous clearing if a sigil or requested clear would do so, anyway.)
For that, we unfortunately need to duplicate the sigil and clear processing into the read loop. Write a separate invokeTargetClear() variant of invokeTargetAddendum() for that.