Rework timeoutReader()
We can't just read the entire remainingLine; characters are produced slowly, this will block until an entire line is ready, and the timer won't be updated.
Instead, we have to read character by character; once a newline has been consumed (recognized by an empty $char), the full line is submitted.