invocationMessage: FIX: Rendering on all output and timeout sometimes loses a line
This seems to happen when a newly read line first goes into the timeout, but the next update already is a new line. The timeout branch reuses the last read line, so we mustn't simply update $line there. Instead, when output is meant to be stored ($isCaptureStdout), remember the $lastLineOnTimeout. Then in the regular branch, if this isn't identical to the last added line ($lastCommandErrorOutputAddedLine - to avoid duplicating output), prepend that, and clear the variable.