Add PROGRESSNOTIFICATION_CLEAR_COMMANDLINE and tests for --to command
Depending on the command, we might need explicit clearing at the end or not. PROGRESSNOTIFICATION_POST_COMMANDLINE is invoked unconditionally, also if there's no progress at all, so it's not optimal for this. So let's introduce a separate PROGRESSNOTIFICATION_CLEAR_COMMANDLINE with whom the user can control clearing, either by supplying a different command-line or turning off explicit clear invocation by emptying the variable.