progressNotification: CHG: Split --count and --percentage into --{ap,pre}pend-* and add --count-to
There's not enough control over the placement; I've tried to make the default looking good by adding special defaults when both are used, but especially with reports of varying size, it might be better to prepend the (rather fixed-width) count and/or percentage. The most universal solution would be printf-like format specifiers, but then clients would have to add them, whereas the report source likely is third-party, so another pipeline step would be needed (<report-source> | prefix '%p: ' | progressNotification ...). Even worse, even client that don't use these would be required to do escaping. And the end count would still have to be passed on.
So, let's just add --prepend-* variants.

Second, a N/M count might be useful as well. As --*-count already takes the WHAT (and unlike the percentages, the end count is optional here), pass this via a separate new --count-to parameter.